JAVA-3485: add ServerDescriptionChangeEventHelperSpecification test
